% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augere.screen-package.R
\docType{package}
\name{augere.screen-package}
\alias{augere.screen}
\alias{augere.screen-package}
\title{augere.screen: Automatic Generation of Functional Screen Analyses}
\description{
Implements pipelines for generating functional screen analysis reports in the augere framework. This uses voom to test for differential abundance of barcodes with consolidation into gene-level statistics. Each pipeline function generates a self-contained Rmarkdown report with all of the steps required to reproduce the analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/augere-bioinfo/augere.screen}
  \item Report bugs at \url{https://github.com/augere-bioinfo/augere.screen/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})

}
