% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augere.screen-package.R
\name{augere.screen-package}
\alias{augere.screen-package}
\title{augere.screen}
\description{
Implement functional screen analyses in the augere framework.
The \code{\link{runScreen}} function tests for differential barcode abundance with voom,
also generating a self-contained Rmarkdown report to reproduce each analysis.
}
